requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.8
autoflake==2.3.1
autopep8==2.3.1
billiard==4.2.1
black==24.10.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
common==0.1.2
crashtest==0.4.1
cron-descriptor==1.4.5
cryptography==44.0.2
dill==0.3.9
distlib==0.3.9
Django==5.1.5
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-codemod==2.3.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-timezone-field==7.1
django-upgrade==1.22.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.8
dulwich==0.22.7
fastjsonschema==2.21.1
filelock==3.16.1
flake8==7.1.1
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
identify==2.6.5
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
kombu==5.4.2
libcst==1.6.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.2.1
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==2.0.1
poetry-core==2.0.1
pre_commit==4.0.1
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.9.0
pylint==3.3.3
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pyproject_hooks==1.2.0
pytest==8.3.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pyupgrade==3.19.1
PyYAML==6.0.2
RapidFuzz==3.11.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-click==1.8.5
rsa==4.9
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.3
tokenize_rt==6.1.0
tomlkit==0.13.2
trove-classifiers==2025.1.10.15
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.28.1
wcwidth==0.2.13
Werkzeug==3.1.3
xattr==1.1.4
